Updated the test case, ensured test passage.
The test case needed to be updated for the latest RSpec version.
The tests were failing because Future and Object didn't remove their
methods if they inherited BasicObject. In that case, equal? wasn't
overridden as well as the equality operators and any RSpec test that
compared the promise result to another result using '==' would fail.