Added thread pool support. Reduced code (but changed API)
Thread pool support has been added by way of a new method:
Promising::promise_with_worker which returns a promise and a worker block
that can be added to a queue processed by a thread pool. The worker block
can also safely be ignored, in which case the returned promise behaves like a
promise.

The Future class has been replaced with a utility method that wraps a newly
executing thread inside a promise, which when forced, waits for, and returns
the thread's result.

The extension of 'Kernel' was replaced with a new module, 'Promising'.

All of this changes the API, but doesn't remove core functionality.

Tests were added for the new behavior and for some behavior that wasn't
already tested.