Adds __worker__ method for thread-pool support
Adds a single method: __worker__ which returns a Proc that can be
placed on the queue of a thread pool.

The method which executes the @block has been moved to a private method
__call_block__. This is so it can be called by both the worker and the
promise. The worker and the promise both share a mutex lock but the
worker uses Mutex#try_lock to obtain it which means if the lock is
obtained, the worker will consider the block called and exit.

Documentation and tests included.